Fallback to shared_store when audio_data is empty
