+++ app/music_assistant_api/ma_routes.py
@@ -1,47 +1,57 @@
-"""Route definitions for music_assistant_api (ma_routes).
-
-This module registers the HTTP endpoints on a provided Flask
-`Blueprint`. It stores the last pushed stream metadata in a
-module-level `_store` variable.
-"""
+"""Route definitions for music_assistant_api (ma_routes)."""
 
 from flask import jsonify, request
+import os
 import time
+from urllib.parse import urlparse, urlunparse
+import shared_store
 
 
-_store = None
-_version = 0  # Increment each time new data is pushed
+def _rewrite_url(url: str) -> str:
+    """Rewrite internal Music Assistant URLs to public hostname."""
+    if not url:
+        return url
+    ma_hostname = os.environ.get('MA_HOSTNAME', '').strip()
+    if not ma_hostname:
+        return url
+    try:
+        parsed = urlparse(url)
+        if not parsed.hostname:
+            return url
+        rewritten = urlunparse((
+            'https', ma_hostname, parsed.path,
+            parsed.params, parsed.query, parsed.fragment
+        ))
+        return rewritten
+    except Exception:
+        return url
 
 
 def register_routes(bp):
     @bp.route('/push-url', methods=['POST'])
     def push_url():
-        """Accept JSON with streamUrl and optional metadata and store it.
-
-        Expected JSON body: { streamUrl, title, artist, album, imageUrl }
-        """
-        global _store, _version
         data = request.get_json(silent=True) or {}
         stream_url = data.get('streamUrl')
         if not stream_url:
             return jsonify({'error': 'Missing required fields'}), 400
 
-        _version += 1
-        _store = {
+        stream_url = _rewrite_url(stream_url)
+        image_url = _rewrite_url(data.get('imageUrl'))
+
+        shared_store._version += 1
+        shared_store._store = {
             'streamUrl': stream_url,
             'title': data.get('title'),
             'artist': data.get('artist'),
             'album': data.get('album'),
-            'imageUrl': data.get('imageUrl'),
-            'version': _version,
+            'imageUrl': image_url,
+            'version': shared_store._version,
             'timestamp': time.time()
         }
-        return jsonify({'status': 'ok', 'version': _version})
+        return jsonify({'status': 'ok', 'version': shared_store._version})
 
     @bp.route('/latest-url', methods=['GET'])
     def latest_url():
-        """Return the last pushed stream metadata for the Music Assistant.
-        """
-        if not _store:
-            return jsonify({'error': 'No URL available, please check if Music Assistant has pushed a URL to the API'}), 404
-        return jsonify(_store)
+        if not shared_store._store:
+            return jsonify({'error': 'No URL available'}), 404
+        return jsonify(shared_store._store)
