+++ app/shared_store.py
@@ -1,0 +1,4 @@
+"""Shared store for MA and Alexa routes."""
+
+_store = None
+_version = 0
